add ui registration completion
