+++ server/storage/registration.ts
@@ -1,0 +1,5 @@
+import { Registration } from "../types/registration";
+
+export default function useRegistrationStorage() {
+  return useStorage<Registration>("registration");
+}
