+++ server/types/registration.ts
@@ -1,0 +1,3 @@
+export type Registration = {
+  token: string;
+};
